# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Cue API
  version: 0.1.0
  description: |
    HTTP contract between cue-api and cue-ios. Source of truth for endpoints,
    request/response shapes, and status codes. See `docs/architecture.md` and
    `docs/specs/` for design rationale.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.cue.app
    description: Production (placeholder — not yet deployed)
...
paths:
...
  /tasks/{id}:
    get:
      tags: [tasks]
      summary: Get a task series row
      description: |
        Returns the full `TaskDTO` for the series anchor, including the embedded
        `recurrence` rule ready for an edit form. Use this to open the detail /
        edit screen.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task id.
      responses:
        '200':
          description: Task series row.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: Task belongs to another user.
        '404':
          description: No task with that id.
    patch:
      tags: [tasks]
      summary: Update a task (all / one-off scope)
      description: |
        Updates the task's fields in the "all" scope — the whole series or the
        one-off task. Omitted fields are left unchanged. `recurrence: null` clears
        the inline config; a `RecurrenceRuleInput` replaces or adds it. `color` and
        `requiresCompletion` may likewise be set or cleared (null re-inherits the
        group). Returns the updated `TaskDTO`.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Updated task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: Task belongs to another user.
        '404':
          description: No task with that id.
    delete:
      tags: [tasks]
      summary: Delete a task
      description: |
        Soft-deletes the task (and its series, for recurring tasks). Returns
        `{ id }` with HTTP 200 — NOT 204, so the iOS client can decode the body.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task id.
      responses:
        '200':
          description: Task deleted.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: string
                    format: uuid
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: Task belongs to another user.
        '404':
          description: No task with that id.
...
  /task-groups:
    get:
      tags: [task-groups]
      summary: List task groups
      description: |
        Returns all task groups for the authenticated user. When `calendarId` is
        provided, scopes the result to that calendar (ownership asserted); otherwise
        returns groups across all calendars the user owns. Each group includes its
        embedded `recurrence` config (or null).
      security:
        - bearerAuth: []
      parameters:
        - name: calendarId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: When given, scopes to a single calendar.
      responses:
        '200':
          description: Task groups.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskGroupDTO'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: The calendar is owned by another user.
        '404':
          description: The calendar does not exist.
    post:
      tags: [task-groups]
      summary: Create a task group
      description: |
        Creates a task group in the referenced calendar. When `recurrence` is
        provided, it is stored inline as the group's default `recurrenceConfig`
        (JSONB — no separate row, ADR 0054); tasks in the group with no own config
        inherit this schedule. Returns the created `TaskGroupDTO`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskGroupRequest'
      responses:
        '201':
          description: Task group created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskGroupDTO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: The calendar is owned by another user.
        '404':
          description: The calendar does not exist.
  /task-groups/reorder:
    patch:
      tags: [task-groups]
      summary: Bulk-reorder task groups
      description: |
        Transactionally reorders the caller's task groups: each group in `groupIds`
        receives a `sortOrder` equal to its index, replacing N racy single PATCHes.
        Every id must exist (else 404) and belong to a calendar the caller owns
        (else 403) — validated BEFORE any mutation, so a partial reorder can never
        persist. Returns the reordered groups in the requested order.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderTaskGroupsRequest'
      responses:
        '200':
          description: The reordered task groups, in the requested order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskGroupDTO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: A referenced group belongs to a calendar the caller does not own.
        '404':
          description: A referenced group id does not exist.
  /task-groups/{id}:
    patch:
      tags: [task-groups]
      summary: Update a task group
      description: |
        Updates the task group's mutable fields. `recurrence: null` clears the
        inline default config; a `RecurrenceRuleInput` replaces or adds it. `color`
        and `requiresCompletion` (the defaults tasks inherit) may likewise be set or
        cleared. Returns the updated `TaskGroupDTO`.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task group id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskGroupRequest'
      responses:
        '200':
          description: Updated task group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskGroupDTO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: Task group belongs to another user.
        '404':
          description: No task group with that id.
    delete:
      tags: [task-groups]
      summary: Delete a task group
      description: |
        Deletes the task group. Contained tasks are not deleted; their `groupId`
        is set to null. Returns `{ id }` with HTTP 200.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task group id.
      responses:
        '200':
          description: Task group deleted.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: string
                    format: uuid
        '401':
          description: Missing, invalid, or expired bearer token.
        '403':
          description: Task group belongs to another user.
        '404':
          description: No task group with that id.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RecurrenceConfig:
      type: object
      description: |
        The inline recurrence config returned inside TaskDTO and TaskGroupDTO.
        Stored as a JSONB column on `task` / `task_group` (no separate row, no
        `id`) — ADR 0054. Field names are normative — the FE decodes them exactly.
      required: [frequency, interval, byWeekday, byMonthDay, byMonth, bySetPos, monthlyAnchor, endType, endDate, count]
      properties:
        frequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, YEARLY]
        interval:
          type: integer
          minimum: 1
          description: Repeat every N units. 1 = every period.
        byWeekday:
          type: ['array', 'null']
          items:
            type: integer
            minimum: 0
            maximum: 6
          description: Weekday ordinals (0 = Mon … 6 = Sun). Null means no restriction.
        byMonthDay:
          type: ['array', 'null']
          items:
            type: integer
            minimum: 1
            maximum: 31
          description: Days of the month. Null means no restriction.
        byMonth:
          type: ['array', 'null']
          items:
            type: integer
            minimum: 1
            maximum: 12
          description: Months (1–12). Null means no restriction.
        bySetPos:
          type: ['array', 'null']
          items:
            type: integer
            enum: [1, 2, 3, 4, -1]
          description: |
            MONTHLY ordinal selectors paired with `byWeekday` — e.g. `[1]` + Monday =
            "first Monday", `[-1]` + Friday = "last Friday". 1..4 or -1. Null means
            no nth-weekday restriction.
        monthlyAnchor:
          type: ['string', 'null']
          enum: [FIRST_WORKDAY, LAST_WORKDAY, DAY_BEFORE_LAST_WORKDAY, null]
          description: |
            MONTHLY working-day anchor — resolves to the first / last / day-before-last
            workday of each month. Mutually exclusive with the `bySetPos`/`byMonthDay`
            axes. Null when no working-day anchor applies.
        endType:
          type: string
          enum: [NEVER, UNTIL_DATE, COUNT]
        endDate:
          type: [string, 'null']
          format: date
          description: Required when endType = UNTIL_DATE.
        count:
          type: [integer, 'null']
          minimum: 1
          description: Required when endType = COUNT.
    RecurrenceRuleInput:
      type: object
      description: |
        Recurrence config payload sent inside create/update task and task-group
        request bodies — the same inline shape as RecurrenceConfig (normalized and
        stored verbatim as JSONB). Cross-field: COUNT requires `count`; UNTIL_DATE
        requires `endDate`.
      required: [frequency]
      properties:
        frequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, YEARLY]
        interval:
          type: integer
          minimum: 1
          default: 1
        byWeekday:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
        byMonthDay:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 31
        byMonth:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 12
        bySetPos:
          type: array
          items:
            type: integer
            enum: [1, 2, 3, 4, -1]
          description: |
            MONTHLY ordinal selectors paired with `byWeekday` ("first Monday",
            "last Friday"). Values 1..4 or -1.
        monthlyAnchor:
          type: string
          enum: [FIRST_WORKDAY, LAST_WORKDAY, DAY_BEFORE_LAST_WORKDAY]
          description: |
            MONTHLY working-day anchor (first / last / day-before-last workday).
            Mutually exclusive with the `bySetPos`/`byMonthDay` axes.
        endType:
          type: string
          enum: [NEVER, UNTIL_DATE, COUNT]
          default: NEVER
        endDate:
          type: string
          format: date
          description: Required when endType = UNTIL_DATE.
        count:
          type: integer
          minimum: 1
          description: Required when endType = COUNT.
    TaskDTO:
      type: object
      description: |
        A task series row. Returned by POST /tasks, PATCH /tasks/:id, GET /tasks/:id.
        Includes the embedded recurrence config for the edit form.
      required:
        [id, calendarId, groupId, title, notes, startAt, endAt, isAllDay, timezone,
         requiresCompletion, color, icon, reminders, completedAt, recurrence,
         notificationStrategyId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        calendarId:
          type: string
          format: uuid
        groupId:
          type: [string, 'null']
          format: uuid
        title:
          type: string
        notes:
          type: [string, 'null']
        icon:
          type: [string, 'null']
          description: Per-task icon (e.g. an SF Symbol name); null when iconless. (S4)
        reminders:
          type: array
          items:
            $ref: '#/components/schemas/ReminderResultDTO'
          description: |
            The task's own per-task reminders (offset + channel), earliest first.
            Empty when the task carries no per-task reminders. (S1)
        startAt:
          type: [string, 'null']
          format: date-time
          description: ISO-8601 start; null for a timeless todo.
        endAt:
          type: [string, 'null']
          format: date-time
        isAllDay:
          type: boolean
        timezone:
          type: string
          description: IANA zone — the anchor for recurrence/DST expansion.
        requiresCompletion:
          type: [boolean, 'null']
          description: |
            Per-task override; null inherits the group, then defaults to false.
            See the effective-settings resolver (ADR 0054, task-wins).
        color:
          type: [string, 'null']
          description: |
            A TaskColor preset name OR a #RRGGBB hex; null inherits the group.
            Effective color resolves task-wins (ADR 0054).
        completedAt:
          type: [string, 'null']
          format: date-time
        recurrence:
          oneOf:
            - $ref: '#/components/schemas/RecurrenceConfig'
            - type: 'null'
          description: |
            Inline recurrence config (JSONB) for the edit form; null when not
            recurring. ADR 0054 — no separate rule id.
        notificationStrategyId:
          type: [string, 'null']
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    TaskGroupDTO:
      type: object
      description: |
        A task group. Returned by all task-group endpoints.
        The embedded `recurrence` is the group default config inherited by tasks
        with no own recurrence; `color` / `requiresCompletion` are likewise the
        group-level defaults a task inherits when it sets none (ADR 0054).
      required:
        [id, calendarId, name, color, requiresCompletion, icon, sortOrder,
         recurrence, defaultNotificationStrategyId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        calendarId:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: [string, 'null']
          description: |
            Group default color — a TaskColor preset name OR a #RRGGBB hex; tasks
            inherit it when they set no own color (ADR 0054).
        requiresCompletion:
          type: [boolean, 'null']
          description: |
            Group default completion requirement; tasks inherit it when they set
            none, falling back to false (ADR 0054 resolver). Nullable, no DB default.
        icon:
          type: [string, 'null']
        sortOrder:
          type: integer
          minimum: 0
        recurrence:
          oneOf:
            - $ref: '#/components/schemas/RecurrenceConfig'
            - type: 'null'
          description: |
            Inline group default recurrence config (JSONB); null when the group has
            no default. ADR 0054 — no separate rule id.
        defaultNotificationStrategyId:
          type: [string, 'null']
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateTaskRequest:
      type: object
      description: All fields optional — omitted keys leave the current value unchanged.
      properties:
        title:
          type: string
          minLength: 1
        notes:
          type: [string, 'null']
        startAt:
          type: [string, 'null']
          format: date-time
        endAt:
          type: [string, 'null']
          format: date-time
        isAllDay:
          type: boolean
        requiresCompletion:
          type: [boolean, 'null']
          description: null clears the per-task override (re-inherit group, then false). ADR 0054.
        color:
          type: [string, 'null']
          description: |
            A TaskColor preset name OR a #RRGGBB hex; null clears the per-task color
            (re-inherit group). ADR 0054.
        icon:
          type: [string, 'null']
          description: Per-task icon (e.g. an SF Symbol name); null clears it. (S4)
        reminders:
          type: array
          items:
            $ref: '#/components/schemas/ReminderRequest'
          description: |
            When provided, REPLACES the full per-task reminder set; an empty array
            clears all reminders; omitting the key leaves the existing set intact. (S1)
        groupId:
          type: [string, 'null']
          format: uuid
          description: Set to null to ungroup.
        recurrence:
          oneOf:
            - $ref: '#/components/schemas/RecurrenceRuleInput'
            - type: 'null'
          description: null removes the inline config; an Input replaces/adds it.
...
    CreateTaskGroupRequest:
      type: object
      required: [calendarId, name]
      properties:
        calendarId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        color:
          type: string
          description: Group default color — a TaskColor preset name OR a #RRGGBB hex.
        requiresCompletion:
          type: [boolean, 'null']
          description: Group default completion requirement inherited by tasks (ADR 0054). No column default.
        icon:
          type: string
        sortOrder:
          type: integer
          minimum: 0
          default: 0
        defaultNotificationStrategyId:
          type: string
          format: uuid
        recurrence:
          $ref: '#/components/schemas/RecurrenceRuleInput'
    UpdateTaskGroupRequest:
      type: object
      description: All fields optional — omitted keys leave the current value unchanged.
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: [string, 'null']
          description: |
            Group default color — a TaskColor preset name OR a #RRGGBB hex; null
            clears it. (Also settable by the assistant via the `update_group` AI tool — ADR 0054.)
        requiresCompletion:
          type: [boolean, 'null']
          description: Group default completion requirement; null clears it. ADR 0054.
        icon:
          type: [string, 'null']
        sortOrder:
          type: integer
          minimum: 0
        recurrence:
          oneOf:
            - $ref: '#/components/schemas/RecurrenceRuleInput'
            - type: 'null'
          description: null removes the group default config; an Input replaces/adds it.
...
    ReminderRequest:
      type: object
      description: A single per-task reminder on a create/update task payload (S1).
      required: [offsetMinutes, channel]
      properties:
        offsetMinutes:
          type: integer
          minimum: -10080
          maximum: 10080
          example: -15
          description: |
            Offset in minutes from the task start (negative fires before, positive
            after), bounded to ±1 week.
        channel:
          type: string
          enum: [PUSH, TELEGRAM]
          description: Delivery channel for this reminder.
    ReminderResultDTO:
      type: object
      description: |
        A per-task reminder returned on TaskDTO (S1). Carries the persisted rule id
        so the client can correlate, plus the offset + channel.
      required: [id, offsetMinutes, channel]
      properties:
        id:
          type: string
          format: uuid
        offsetMinutes:
          type: integer
          example: -15
        channel:
          type: string
          enum: [PUSH, TELEGRAM]
...
    ReorderTaskGroupsRequest:
      type: object
      description: |
        Bulk-reorder payload (the complete, ordered list of group ids). Each group
        receives a `sortOrder` equal to its index, transactionally. Every id must
        exist (else 404) and belong to a calendar the caller owns (else 403).
      required: [groupIds]
      properties:
        groupIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            format: uuid
          description: Ordered group ids; each row receives a sortOrder equal to its index.
...
    ValidationErrorBody:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
          example: Bad Request
  responses:
    ValidationError:
      description: Request DTO failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorBody'
...